third_party/blink/web_tests/flag-specific/disable-site-isolation-trials/external/wpt/web-share/disabled-by-permissions-policy-cross-origin.https.sub-expected.txt

This is a testharness.js-based test.
PASS share() is disabled by default 'self' by permissions policy for cross-origin iframes
PASS share() is disabled explicitly by permissions policy for cross-origin iframe
PASS share() not allowed, as only allowed to share with self
PASS canShare() not allowed to share by default permissions policy cross-origin
FAIL canShare() is allowed by permissions policy to share cross-origin on a particular origin assert_equals: Expected true, is it can now share on https://web-platform.test:8444. expected true but got false
PASS canShare() with self
Harness: the test ran to completion.
